Possibility to clone into imagefile

+++ sbin/clone_sd.pl
@@ -5,6 +5,7 @@ use JSON;
 use LoxBerry::System;
 use LoxBerry::Log;
 use Data::Dumper;
+use LoxBerry::Storage;
 
 my $version = "2.2.0.3";
 
@@ -27,6 +28,7 @@ my $log = LoxBerry::Log->new (
 LOGSTART "Clone SD card";
 LOGINF "Version of this script: $version";
 my $destpath = $ARGV[0];
+my $desttype = $ARGV[1];
 
 my $curruser = getpwuid($<);
 LOGINF "Executing user is $curruser";
@@ -105,8 +107,9 @@ foreach my $blockdevice ( @{$lsblk->{blockdevices}} ) {
 	}
 }
 print "\n";
-if (!$destpath) {
-	print "Your possible clone devices (DESTINATION of clone):\n";
+if ( !$destpath || !$desttype || ($desttype ne "device" && $desttype ne "path") ) {
+	print "You can clone directly onto a new device (like a SDCard) or into an imagefile saved into an external path (usb or netshare), which then has do be flashed onto a SDcard.\n\n";
+	print "Your possible clone \033[1mdevices\033[0m (DESTINATION of clone, type device):\n";
 	foreach my $blockdevice ( @{$lsblk->{blockdevices}} ) {
 		next if ($blockdevice->{isboot});
 		print "/dev/$blockdevice->{name} Type $blockdevice->{type} Size " . LoxBerry::System::bytes_humanreadable($blockdevice->{size}) ."\n";
@@ -115,12 +118,23 @@ if (!$destpath) {
 			print "   -> Partition $partition->{name} Size " . LoxBerry::System::bytes_humanreadable($partition->{size}) ." " . uc($partition->{fstype}) . "\n";;
 		}
 	}
-	print "\nThe process will COMPLETELY DELETE your selected destination device. Mind your step.\n";
+	print "\nIf you use type device as destination, the process will \033[1mCOMPLETELY DELETE\033[0m your selected destination device. Double check before starting!\n";
 	print "\n";
-	print "To start the clone, use the DESTINATION device as parameter.\n";
-	print "Example: $0 /dev/sdc\n";
+
+	print "Your possible clone \033[1mpathes\033[0m (DESTINATION of clone, type path):\n";
+
+	my @storages = LoxBerry::Storage::get_storage(1);
+	foreach my $path ( @storages ) {
+		next if $path->{TYPE} eq "local";
+		print "$path->{NAME}: $path->{PATH} Type $path->{GROUP}/$path->{TYPE} Available space " . LoxBerry::System::bytes_humanreadable($path->{AVAILABLE}*1024) . "\n";
+	}
+	print "\nIf you use type path as destination, the storage path must have enough free disc space (\033[1mtwice\033[0m as the size of your SDcard!).\n";
+
+	print "\nTo start the clone, use the DESTINATION device/path as first parameter and the device type [device|path] as second parameter.\n";
+	print "Examples: $0 /dev/sdc device\n";
+	print "          $0 /opt/loxberry/system/storage/usb/31816dac-01 path\n";
 	if ( ! -e "$lbsconfigdir/is_raspberry.cfg" ) {
-		print " DANGER!! This not the original LoxBerry image (not on a Raspberry).\n";
+		print "\n \033[1mDANGER!!\033[0m This not the original LoxBerry image (not on a Raspberry).\n";
 		print "          No idea what happens on a Non-Raspberry installation.\n";
 	}
 	exit(0);
@@ -130,40 +144,79 @@ if (!$destpath) {
 # User input validation
 ########################
 
-# Check if $destpath exists and is not boot
+my $destdevice;
 my $destpath_found = 0;
 my $destdevice_index = 0;
-foreach my $blockdevice ( @{$lsblk->{blockdevices}} ) {
-	if ( $destpath eq $blockdevice->{path} ) {
-		$destpath_found = 1;
+my $required_space;
+
+# Check if $destpath exists and is not boot
+if ( $desttype eq "device" ) {
+	foreach my $blockdevice ( @{$lsblk->{blockdevices}} ) {
+		if ( $destpath eq $blockdevice->{path} ) {
+			$destpath_found = 1;
 		
-		if( $blockdevice->{isboot} ) {
-			LOGCRIT "Your entered DESTINATION path $destpath is the boot device!";
-			exit(1);
+			if( $blockdevice->{isboot} ) {
+				LOGCRIT "Your entered DESTINATION path $destpath is the boot device!";
+				exit(1);
+			}
+			last;
 		}
-		last;
+		$destdevice_index++;
 	}
-	$destdevice_index++;
+	if (!$destpath_found) {
+		LOGCRIT "Your entered DESTINATION path $destpath does not exist.";
+		exit(1);
+	}
+
+	# For easiness, set $destdevice to the destination sevice of the lsblk array
+	$destdevice = $lsblk->{blockdevices}[$destdevice_index];
+
+	print "Destination device is $destdevice->{name} ($destdevice->{path}) size " . LoxBerry::System::bytes_humanreadable($destdevice->{size}) . "\n";
+
+	# Size check (used size + $dest_bootpart_size plus 20%)
+	$required_space = ($src1_used + $dest_bootpart_size*1024*1024)*1.2;
+	if( $destdevice->{size} < $required_space ) {
+		LOGCRIT "$destpath (" . LoxBerry::System::bytes_humanreadable($destdevice->{size}) . ") is smaller that required space (" . LoxBerry::System::bytes_humanreadable($required_space) . ")";
+		exit(1);
+	}
 }
-if (!$destpath_found) {
-	LOGCRIT "Your entered DESTINATION path $destpath does not exist.";
-	exit(1);
+
+# Check if $destpath exists and is not local
+if ( $desttype eq "path" ) {
+	my $destpath_found = 0;
+	my $destdevice_index = 0;
+	my @storages = LoxBerry::Storage::get_storage(1);
+	foreach my $path ( @storages ) {
+		$destdevice_index++;
+		next if $path->{TYPE} eq "local";
+		if ( $destpath eq $path->{PATH} ) { 
+			$destpath_found = 1;
+			$destdevice_index--;
+			last;
+		}
+	}
+
+	if (!$destpath_found) {
+		LOGCRIT "Your entered DESTINATION path $destpath does not exist or isn't writeable.";
+		exit(1);
+	}
+
+	# For easiness, set $destdevice to the destination sevice of the lsblk array
+	$destdevice = $storages[$destdevice_index];
+
+	print "Destination device is $destdevice->{NAME} ($destdevice->{PATH}) available space " . LoxBerry::System::bytes_humanreadable($destdevice->{AVAILABLE}*1024) . "\n";
+
+	# Size check (used size + $dest_bootpart_size plus 20%)
+	$required_space = ($src1_used + $dest_bootpart_size*1024*1024)*1.2;
+	if( $destdevice->{AVAILABLE}*1024 < $required_space ) {
+		LOGCRIT "$destpath (" . LoxBerry::System::bytes_humanreadable($destdevice->{AVAILABLE}*1024) . ") is smaller that required space (" . LoxBerry::System::bytes_humanreadable($required_space) . ")";
+		exit(1);
+	}
 }
 
-# For easiness, set $destdevice to the destination sevice of the lsblk array
-my $destdevice = $lsblk->{blockdevices}[$destdevice_index];
-
-print "Destination device is $destdevice->{name} ($destdevice->{path}) size " . LoxBerry::System::bytes_humanreadable($destdevice->{size}) . "\n";
-
-# Size check (used size + $dest_bootpart_size plus 20%)
-my $required_space = ($src1_used + $dest_bootpart_size*1024*1024)*1.2;
-if( $destdevice->{size} < $required_space ) {
-	LOGCRIT "$destpath (" . LoxBerry::System::bytes_humanreadable($destdevice->{size}) . ") is smaller that required space (" . LoxBerry::System::bytes_humanreadable($required_space) . ")";
-	exit(1);
-}
-
-print "\nPress Ctrl-C now, if you have changed your mind\n";
+print "\n\033[1mPress Ctrl-C now, if you have changed your mind\033[0m (I will wait 5sec and then continue automatically).\n";
 sleep(5);
+
 ##################################
 # Start
 ##################################
@@ -173,14 +226,15 @@ print "Too late - let the game begin!\n\n";
 execute( command => "systemctl stop autofs", log => $log );
 
 # Unmount mounted destination partitions
-
-for my $i ( 1..3) {
-	LOGINF "Unmount Try $i";
-	foreach my $partition ( @{$destdevice->{children}} ) {
-		LOGDEB "umount ".$partition->{path};
-		my ($rc) = execute( command => "umount -q -A ".$partition->{path} );
+if ($desttype eq "device") {
+	for my $i ( 1..3) {
+		LOGINF "Unmount Try $i";
+		foreach my $partition ( @{$destdevice->{children}} ) {
+			LOGDEB "umount ".$partition->{path};
+			my ($rc) = execute( command => "umount -q -A ".$partition->{path} );
+		}
+		sleep 1;
 	}
-	sleep 1;
 }
 
 my @helperdirs = (
@@ -208,31 +262,41 @@ foreach my $helperdir ( @helperdirs ) {
 	}
 }
 
-LOGINF "Deleting old partitions on your destination";
-execute( command => "wipefs -a $destpath", log => $log );
-sleep(1);
+if ($desttype eq "device") {
+	LOGINF "Deleting old partitions on your destination";
+	execute( command => "wipefs -a $destpath", log => $log );
+	sleep(1);
 
-# Re-read lsblk
-LOGINF "Re-checking destination sd card";
-my $checklsblk = lsblk();
-$destpath_found = 0;
-$destdevice_index = 0;
+	# Re-read lsblk
+	LOGINF "Re-checking destination sd card";
+	my $checklsblk = lsblk();
+	$destpath_found = 0;
+	$destdevice_index = 0;
 
-foreach my $blkdevice ( @{$checklsblk->{blockdevices}} ) {
-	if( $blkdevice->{path} eq $destpath ) {
-		$destpath_found = 1;
-		last;
+	foreach my $blkdevice ( @{$checklsblk->{blockdevices}} ) {
+		if( $blkdevice->{path} eq $destpath ) {
+			$destpath_found = 1;
+			last;
+		}
+		$destdevice_index++;
 	}
-	$destdevice_index++;
-}
-if ( !$destpath_found ) {
-	LOGCRIT "Could not find $destpath anymore after wiping partitions.";
-	exit(1);
+	if ( !$destpath_found ) {
+		LOGCRIT "Could not find $destpath anymore after wiping partitions.";
+		exit(1);
+	}
+
+	if ( defined $checklsblk->{blockdevices}[$destdevice_index]->{children} and scalar @{$checklsblk->{blockdevices}[$destdevice_index]->{children}} != 0 ) {
+		LOGCRIT "Not all partitons could be deleted on destination $destpath.";
+		exit(1);
+	}
 }
 
-if ( defined $checklsblk->{blockdevices}[$destdevice_index]->{children} and scalar @{$checklsblk->{blockdevices}[$destdevice_index]->{children}} != 0 ) {
-	LOGCRIT "Not all partitons could be deleted on destination $destpath.";
-	exit(1);
+if ($desttype eq "path") {
+	LOGINF "Creating destination image file, size " . LoxBerry::System::bytes_humanreadable($required_space);
+	$destpath = $destpath . "/" . LoxBerry::System::lbhostname() . "_image.img";
+	my $targetsize_mb = int( ($required_space / 1024 / 1024) + 0.5 ); # rounded
+	print "Targetsize is: $targetsize_mb\n";
+	execute ( command => "dd if=/dev/zero of=$destpath bs=1 count=0 seek=" . $targetsize_mb . "MB", log => $log );
 }
 
 LOGINF "Creating partiton table";
@@ -245,19 +309,37 @@ execute( command => "parted -s $destpath mkpart primary ext4 ". $dest_bootpart_s
 
 sleep (1);
 
-my $destpath1 = $destpath."1";
-my $destpath2 = $destpath."2";
+my $destpath1;
+my $destpath2;
+my $loop;
 
 # If partitions got re-mounted, unmount again
 # Unmount mounted destination partitions
-
-for my $i ( 1..3 ) {
-	LOGINF "Unmount Try $i";
-	foreach my $partition ( @{$destdevice->{children}} ) {
-		LOGDEB "umount ".$partition->{path};
-		my ($rc) = execute( command => "umount -q -A ".$partition->{path} );
+if ($desttype eq "device") {
+	$destpath1 = $destpath."1";
+	$destpath2 = $destpath."2";
+	for my $i ( 1..3 ) {
+		LOGINF "Unmount Try $i";
+		foreach my $partition ( @{$destdevice->{children}} ) {
+			LOGDEB "umount ".$partition->{path};
+			my ($rc) = execute( command => "umount -q -A ".$partition->{path} );
+		}
+		sleep 1;
 	}
-	sleep 1;
+}
+
+# Create devices fronm image file for each partition
+if ($desttype eq "path") {
+	LOGINF "Create loop devices for all partitions in image file $destpath";
+	execute( command => "losetup -f --show -P $destpath", log => $log );
+	$loop = execute( command => "losetup -j $destpath | cut -d: -f1" );
+	chomp ($loop);
+	if ( $loop !~ /^\/dev\/loop\d+/ ) {
+		LOGCRIT "Could not create loop device.";
+		exit (1);
+	}
+	$destpath1 = $loop."p1";
+	$destpath2 = $loop."p2";
 }
 
 LOGINF "Formatting fat32 boot partition (takes a second)";
@@ -268,13 +350,15 @@ execute( command => "mkfs.ext4 ".$destpath2, log => $log );
 # If partitions got re-mounted, unmount again
 # Unmount mounted destination partitions
 
-for my $i ( 1..3) {
-	LOGINF "Unmount Try $i";
-	foreach my $partition ( @{$destdevice->{children}} ) {
-		LOGDEB "umount ".$partition->{path};
-		my ($rc) = execute( command => "umount -q -A ".$partition->{path} );
+if ($desttype eq "device") {
+	for my $i ( 1..3) {
+		LOGINF "Unmount Try $i";
+		foreach my $partition ( @{$destdevice->{children}} ) {
+			LOGDEB "umount ".$partition->{path};
+			my ($rc) = execute( command => "umount -q -A ".$partition->{path} );
+		}
+		sleep 1;
 	}
-	sleep 1;
 }
 
 # Mounting and checking all mounts
@@ -312,12 +396,30 @@ if (!$mountsok) {
 	exit(1);
 }
 
-
 LOGINF "Copy data of /boot partition (this will take some seconds)";
 execute( command => "cd /media/src1 && tar cSp --numeric-owner --warning='no-file-ignored' -f - . | (cd /media/dst1 && tar xSpvf - )", log => $log );
 LOGINF "Copy data of / partition (this may take an hour...)";
 execute( command => "cd /media/src2 && tar cSp --numeric-owner --warning='no-file-ignored' -f - . | (cd /media/dst2 && tar xSpvf - )", log => $log );
 
+# Delete not needed swap file - this is normally quite big
+if (-e "/media/dst2/var/swap") {
+	unlink ("/media/dst2/var/swap");
+}
+
+# if we clone into an imagefile, we want to resize the SDcard after first booting
+if ($desttype eq "path") {
+	LOGINF "Resize SDcard after first boot again";
+	unlink ("/media/dst1/rootfsresized");
+	open(F,">/media/dst2$lbhomedir/system/daemons/system/98-resizerootfs");
+	print F <<EOF;
+#!/bin/bash
+$lbssbindir/resize_rootfs > $lbslogdir/rootfsresized.log 2>&1
+rm $lbhomedir/system/daemons/system/98-resizerootfs
+reboot
+EOF
+	close (F);
+	execute ( command => "chmod +x /media/dst2$lbhomedir/system/daemons/system/98-resizerootfs" );
+}
 
 LOGINF "Change the PTUUID of destination card to $src_ptuuid";
 `fdisk $destpath <<EOF > /dev/null
@@ -338,7 +440,7 @@ my $newdestpath_found = 0;
 my $newdestdevice_index = 0;
 my $newdst_ptuuid;
 foreach my $blockdevice ( @{$newlsblk->{blockdevices}} ) {
-	if ( $destpath eq $blockdevice->{path} ) {
+	if ( $destpath eq $blockdevice->{path} || $loop eq $blockdevice->{path} ) {
 		$destpath_found = 1;
 		$newdst_ptuuid = $blockdevice->{ptuuid};
 		last;
@@ -361,6 +463,12 @@ foreach my $mountpoint ( keys %chk_mounts ) {
 	}
 }
 
+# Loop devices gives not back correct partuuid - so overwrite here
+if ($desttype eq "path") {
+	$newdst_ptuuid = execute ( command => "fdisk -l $destpath | grep 'Disk identifier' | cut -d: -f2 | cut -dx -f2");
+	chomp ($newdst_ptuuid);
+}
+
 if (!$newdst_ptuuid or $newdst_ptuuid ne $src_ptuuid) {
 	LOGERR "Source PTUUID is $src_ptuuid, Dest PTUUID is $newdst_ptuuid - They are not equal. The new SD may fail to boot!";
 } else {
@@ -371,7 +479,7 @@ LOGWARN "If it fails to boot, connect a display to Raspberry to check what happe
 LOGEND "Finished";
 
 # Cleaning up
-reboot_required("After clone_sd.pl usage you need to reboot LoxBerry.");
+#reboot_required("After clone_sd.pl usage you need to reboot LoxBerry.");
 foreach my $mountpoint ( keys %chk_mounts ) {
 	rmdir "$destpath2".$mountpoint;
 }
@@ -381,8 +489,9 @@ foreach my $mountpoint ( keys %chk_mounts ) {
 	execute( command => "umount $mountpoint" );
 	rmdir "$mountpoint";
 }
+execute ( command => "losetup -d $loop" ) if ($desttype eq "path");
 
-
+exit (0);
 	
 ##########
 # Subs
